Fixed wrong buffer reading.
If the file/buffer size is less than 4 kB the buffer may contain data from another buffer. If this happens, all subsequent checks will most probably fail. The .fseq files which with we are dealing with are mostly bigger than those 4 kB - thus this issue wouldn't occur. The problem is if the FSEQ Version was not set to "V2 Uncompressed" like documented in the tesla light show repo, but was left to the default of "V2 ZSTD" - then files sizes <4 kB are very much possible. As this is a case the validator shall handle, those small file sizes need to be able to be read.